Move Sample module to tests
